requirements.txt
Django>=1.7
django-appconf>=0.6
django-ipware>=1.1.3